refactor: move login and signup controllers to a folder
