--- src/presentation/controllers/login/login-contrller-protocols.ts
@@ -1,3 +1,0 @@
-export * from '../../protocols'
-export * from '../../../domain/usecases/authentication'
-export * from '../../protocols/validation'
