+++ src/presentation/controllers/login/login/login-contrller-protocols.ts
@@ -1,0 +1,3 @@
+export * from '../../../protocols'
+export * from '../../../../domain/usecases/authentication'
+export * from '../../../protocols/validation'
