--- src/presentation/controllers/signup/signup-controller-protocols.ts
@@ -1,5 +1,0 @@
-export * from '../../protocols'
-export * from '../../../domain/usecases/add-account'
-export * from '../../../domain/models/account'
-export * from '../../protocols/validation'
-export * from '../../../domain/usecases/authentication'
